pyproject.toml
[project]
name = "papermatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "backoff>=2.2.1",
    "fastapi[standard]>=0.115.14",
    "gradio>=5.35.0",
    "mixedbread>=0.17.0",
    "pymilvus>=2.5.11",
    "python-dotenv>=1.1.1",
]
